refactor: drawing 패키지 생성 및 클래스 이동
